Organise classes into one package `duke`
