+++ src/main/java/duke/utils/DateTime.java
@@ -1,0 +1,7 @@
+package duke.utils;
+
+import java.time.format.DateTimeFormatter;
+
+public class DateTime {
+    public static final DateTimeFormatter formatter = DateTimeFormatter.ofPattern("d/M/yyyy HHmm");
+}
