--- src/main/java/utils/DateTime.java
@@ -1,7 +1,0 @@
-package utils;
-
-import java.time.format.DateTimeFormatter;
-
-public class DateTime {
-    public static final DateTimeFormatter formatter = DateTimeFormatter.ofPattern("d/M/yyyy HHmm");
-}
