+++ text-ui-test/EXPECTED.TXT
@@ -10,17 +10,17 @@ Hello from
 ――――――――――――――――――――――――――――――――――――――――――――――――――
 Got it. I've added this task:
 	[T][ ] read book
-Now you have 1 tasks in the list.
+Now you have 1 duke.tasks in the list.
 ――――――――――――――――――――――――――――――――――――――――――――――――――
 ――――――――――――――――――――――――――――――――――――――――――――――――――
 Got it. I've added this task:
 	[D][ ] return book(by: 24/4/2019 1600)
-Now you have 2 tasks in the list.
+Now you have 2 duke.tasks in the list.
 ――――――――――――――――――――――――――――――――――――――――――――――――――
 ――――――――――――――――――――――――――――――――――――――――――――――――――
 Got it. I've added this task:
 	[E][ ] project meeting(at: 24/4/2019 1600)
-Now you have 3 tasks in the list.
+Now you have 3 duke.tasks in the list.
 ――――――――――――――――――――――――――――――――――――――――――――――――――
 ――――――――――――――――――――――――――――――――――――――――――――――――――
 Nice! I've marked this task as done:
